+++ streamsx/standard/_adapter.py
@@ -2,9 +2,10 @@
 # Licensed Materials - Property of IBM
 # Copyright IBM Corp. 2017
 
+import streamsx.spl.op
+from streamsx.topology.schema import StreamSchema
+
 import enum
-
-__all__ = ['Format', 'Compression']
-
 Format = enum.Enum('Format', 'csv txt bin block line')
 Compression = enum.Enum('Compression', 'zlib gzip bzip2')
+
